Input: silead: Add support for capactive home button found on some x86 tablets (WIP)
On some x86 tablets with a silead touchscreen the windows logo on the
front is a capacitive home button. Touching this button results in a touch
with bits 12-15 of the Y coordinates set, while normally only the lower 12
are used.

Detect this and report a KEY_LEFTMETA press when this happens. Note for
now we only respond to the Y coordinate bits 12-15 containing 1, assuming
that they may contain other values on (future?) tablets with multiple
softbuttons.

TODO: check there are no false positives on other tablets, remove debug
pr_err once done.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>