[PATCH] ACPI: Adjust the return value of _REV on x86
Since we aim to be compatible with Windows, this tends to break things.
The ideal solution would be to wait for an _OSI() query matching Windows
and then change behaviour, but Windows responds to _REV with 2 even
before that and so vendors might simply change the order of queries in
order to continue IDing Linux. The easiest thing for us to do is just to
change the value on X86 and leave a comment describing why everything is
so awful.

Signed-off-by: Matthew Garrett <mjg59@srcf.ucam.org>