ASoC: Intel: Kconfig: finer granularity for Skylake+ platforms
The current SST driver handles all platforms from Skylake to CannonLake.
This is nice, but it forces users to add platforms they may not want.
It also creates co-existence issues if e.g. a distro wants to enable SOF
for ApolloLake but not for SKL or wants the legacy driver for a specific
target (note that this would require additional changes to PCI match
tables).

Change the SND_SOC_INTEL_SKYLAKE to become a shortcut to select all
platforms (mainly for backwards compatibility). For users who care,
a submenu will be shown.

FIXME:
Is there a better UI configuration? If SND_SOC_INTEL_SKYLAKE doesn't
have a description after 'tristate' it is removed. Adding it as a default
prevents its removal.
Do we want to split BXT and ApolloLake?
Do we want a separate entry for CoffeeLake?

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>