Patch for ksyms error
Nicholas Piggin - Nov. 4, 2016, 7:10 a.m.

lib-ksyms.o is created by linking an empty input file with a linker
script containing the interesting bits. Currently the empty input file
is an archive containing nothing, however this causes the gold linker
to segfault.

I have opened a bug against gold

  https://sourceware.org/bugzilla/show_bug.cgi?id=20767

However this can be worked around by assembling an empty file to link
with instead. The resulting lib-ksyms.o is slightly larger (seemingly
due to empty .text, .data, .bss setions added), but final linked
output should not be changed.

Reported-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
---

Hi Michal,

If we can sneak this into 4.9 it would be nice, so we don't have a
release version of the kernel out there with this bug.

Thanks,
Nick

 scripts/Makefile.build | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)