ASoC: Intel: sst: Free streams on suspend, re-alloc on resume
Some SST-DSP firmware versions loose track of all streams over a
suspend/resume, failing any attempts to resume and/or free streams,
with a SST_ERR_INVALID_STREAM_ID error.

This commit adds support for free-ing the streams on suspend and
re-allocating them on resume, fixing suspend/resume issues on devices
using these firmware versions.

This new behavior gets triggered by a new flag, for now this flag lives
in snd_soc_acpi_mach, so that we can enable this on a per machine level.
But I have a feeling that we may need to simply enable this on all Bay
Trail devices, at which point sst_platform_info would be a better place.

This has been tested on the following devices:
-Asus T100TA, Bay Trail + ALC5642 ("10EC5640" HID), needs the workaround
-GPD win, Cherry Trail + rt5645, does not need the workaround

Tested-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>