ASoC: rt5651: Simplify set_bias_level()
There is no need to set the LDO voltage to 1.2 volt each time we enter
standby, instead always leave it 1.2 volt on BIAS_OFF. Note we do a
snd_soc_codec_force_bias_level(BIAS_OFF) on probe, so this will configure
it correctly right from the start.

For PWR_ANLG2 leave the RT5651_PWR_JD_M and PLL bits as is instead of
having different code-paths for when we've jack-detect vs when we don't.

Note that this also stops enabling the PLL bit when we've a jd_src and we
are running of the RCCLK, this is intentional.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>