ASoC: rt5651: Allow specifying micbias over-current thresholds through pdata
OVer-Current-Detection (OVCD) for the micbias current is used to detect if
an inserted jack is a headset or headphones (mic shorted to ground).

Some boards may need different values for the OVCD threshold because of a
resistor on the board in serial with or parallel to the jack mic contact.

This commit adds support for the sofar unset OVCD scale-factor register
values and adds support for specifying both the current threshold and the
scale-factor to pdata and this commets sets these values only once from
rt5651_set_jack() instead of setting them every time we do jack-detection.

This commit sets the new pdata values for this to 2000uA with a
scale-factor of 0.75 for the KIANO SlimNote 14.2 device, which is the
only rt5652 using device on which jack-detection is currently enabled.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>