brcmfmac: Fix setting of p2p address
The brcmfmac firmware has a requirement for the p2p-dev interface that it
should be different from the mac address of the primary interface, ie.
wlp1s0 in this log. In brcmfmac we try to do that by setting the local
admin bit, but if it is already set we ended up using the same mac address
and getting an -EBUSY error from the firmware.

This would happen when MAC address of the primary interface was changed
(typically for privacy reasons) to one which already has the local admin
bit set.

This commit fixes this by instead using the permanent ethernet address of
the device when no MAC address is specified when instantiating the p2p
interface.