efi: fix efi_pci_io_protocol32 prototype for mixed mode
Mixed mode allows a kernel built for x86_64 to interact with 32-bit
EFI firmware, but requires us to define all struct definitions carefully
when it comes to pointer sizes. efi_pci_io_protocol32 currently uses a
void* for the 'romimage' field, which will be interpreted as a 64-bit
field on such kernels, potentially resulting in bogus memory references
and subsequent crashes.

Cc: <stable@vger.kernel.org>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>