drm/i915: Add an update callback to intel_encoder and call this on fastsets
When we are doing a fastset (needs_modeset=false, update_pipe=true) we
may need to update some encoder-level things such as checking that psr
is enabled.

This commit adds an update callback to intel_encoder and a new
intel_encoders_update helper which calls this for all encoders connected
to a crtc. The new intel_encoders_update helper is called from
intel_update_crtc this when doing a fastset.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>