platform/x86: i2c-multi-instantiate: Skip I2C resources on -EADDRNOTAVAIL
Some ACPI tables also put the SMBus Alert Response Address (ARA, 0x0c) in
their ACPI resource (_CRS) table; and they put it there as the first entry,
here is an example from the CPLM3218 device in the DSDT of an Asus T100TA:

 Method (_CRS, 0, NotSerialized)  // _CRS: Current Resource Settings
 {
     Name (SBUF, ResourceTemplate ()
     {
         I2cSerialBusV2 (0x000C, ControllerInitiated, 0x00061A80,
             AddressingMode7Bit, "\\_SB.I2C3",
             0x00, ResourceConsumer, , Exclusive,
             )
         I2cSerialBusV2 (0x0048, ControllerInitiated, 0x00061A80,
             AddressingMode7Bit, "\\_SB.I2C3",
             0x00, ResourceConsumer, , Exclusive,
             )
         Interrupt (ResourceConsumer, Level, ActiveHigh, Exclusive, ,, )
         {
             0x00000033,
         }
     })
     Return (SBUF) /* \_SB_.I2C3.ALSD._CRS.SBUF */
 }

The actual I2C address of the sensor in this case is the 0x48 address
from the second resource-table entry.

After recent changes i2c_acpi_new_device() will return -EADDRNOTAVAIL when
the I2C resource points to the ARA, change i2c_multi_inst_probe() to skip
the I2C resource for this error rather then to exit with an error, so that
we do instantiate an i2c-client for the second I2C resource.

Note this will cause i2c_unregister_device() to get called on the returned
ERR_PTR on cleanup. This is ok, i2c_unregister_device() ignores the call
if the passed in client pointer IS_ERR_OR_NULL().

Signed-off-by: Hans de Goede <hdegoede@redhat.com>