Integrate KernelSU from LSposed team
The module was orginally created for gaining direct root on phones, later on got a bringup by LSposed team for gaining root access on Windows Subsystem for Android. We're gonna use it for temporary gaining root access for s-x86 builds.
For more info of the module : https://github.com/LSPosed/WSA-Kernel-SU